End to end working of game

+++ Cowboy Simulator/Assets/Scripts/Join.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class Join
 {
-    public int playerId { get; set; }
+    public long playerId { get; set; }
     public string playerName { get; set; }
 }
